% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_utils.R
\name{git_cleanup_local}
\alias{git_cleanup_local}
\title{Cleanup Local Git Repo}
\usage{
git_cleanup_local()
}
\description{
Prunes branches (removes them from list of remote branches) from your local repo that are removed in the remote (origin).
Deletes branches locally that have status "merged" (removes them from list of local branches).
}
\examples{
\dontrun{
git_cleanup_local()
}

}
